pyproject.toml
[project]
name = "it-helpdesk-bot"
version = "0.1.0"
description = "IT Helpdesk Bot with RAG - Technical Support & Troubleshooting Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.2.1",
    "gradio>=5.49.1",
    "isort>=7.0.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.2",
    "openai>=2.6.1",
    "pydantic-settings>=2.11.0",
    "pypdf>=3.17.0",
    "pymupdf>=1.23.0",
    "markdown>=3.5.0",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
packages=["knowledge_chat"]
